refactor: put all managers in a single package
