+++ src/main/java/app/AppBuilder.java
@@ -6,6 +6,8 @@ import interface_adapter.execute_buy.ExecuteBuyController;
 import interface_adapter.execute_buy.ExecuteBuyPresenter;
 import use_case.execute_buy.ExecuteBuyInputBoundary;
 import use_case.execute_buy.ExecuteBuyInteractor;
+import utility.ServiceManager;
+import utility.ViewManager;
 import view.components.DialogComponent;
 import view.panels.DashboardPanel;
 import view.panels.LogInPanel;
