+++ src/main/java/use_case/execute_buy/ExecuteBuyDataAccessInterface.java
@@ -5,7 +5,6 @@ import entity.User;
 public interface ExecuteBuyDataAccessInterface {
     User getUserWithCredential(String credential) throws ValidationException;
 
-    // TODO: define this exception elsewhere
     class ValidationException extends Exception {
     }
 }
