+++ src/main/java/use_case/execute_buy/ExecuteBuyInputData.java
@@ -1,6 +1,7 @@
 package use_case.execute_buy;
 
 public record ExecuteBuyInputData(
+        String credential,
         String ticker,
         int quantity
 ) {
