+++ src/main/java/utility/SessionManager.java
@@ -1,4 +1,4 @@
-package app;
+package utility;
 
 import java.util.HashMap;
 import java.util.Map;
