+++ src/main/java/view/components/DialogComponent.java
@@ -1,7 +1,7 @@
 package view.components;
 
+import utility.ViewManager;
 import view.IComponent;
-import app.ViewManager;
 import view.view_events.DialogEvent;
 import view.view_events.EventType;
 import view.view_events.ViewEvent;
